move testUtils into __tests__ folder
