--- packages/core/src/testUtils/index.ts
@@ -1,1 +1,0 @@
-export { act } from './act';
